Implement priority scheduling
Insert a thread to `ready_list` ordered by `higher_priority()`, which
returns `true` when the priority of left item is higher than it of right
item, `false` otherwise.

In `thread_unblock()`, it checks whether it should yield to other thread
or not. An external interrupt handler must not sleep or yield and idle
thread handles it automatically, so opt out.

When we call `thread_set_priority()`, the priority order can be changed.
If it did, we also yield immediately.